requirements.txt
djangorestframework
psycopg
psycopg[binary]
requests
